refactor: simplify recents feature — deduplicate groupByTime, extract resolveItemDisplay, remove dead lfm toggle code
Moved groupByTime to useRecentSelections so Search and RecentsPopover share one implementation instead of two. Extracted resolveItemDisplay helper to replace copy-pasted image/color resolution in both popover views. Removed the unused lfmToggle/clearButton variables and all the dead lfm state they depended on (showLfm, lfmList, useLastFmRecentTracks, Switch). Added isMounted guard to useLastFmRecentTracks to prevent setState after unmount.
